clean it up a bit
