+++ server/src/db/bits/bits.methods.ts
@@ -1,0 +1,8 @@
+/**
+ * Instance methods on bit documents.
+ * @packageDocumentation
+ */
+
+ export const methods = {
+
+ };
