--- server/src/db/bits/bits.model.ts
@@ -1,5 +1,0 @@
-import { model } from "mongoose";
-import { BitDocument } from "./bits.types";
-import {BitsSchema} from './bits.schema';
-
-export const BitsModel = model<BitDocument>('Post', BitsSchema);
