+++ server/src/db/bits/bits.schema.ts
@@ -1,7 +1,25 @@
-import { Schema } from "mongoose";
-import { queryBits } from "./bits.statics";
+import { Schema, Document, Model, model } from "mongoose";
+import * as statics from "./bits.statics";
+import * as methods from './bits.methods';
 
-export const BitsSchema = new Schema({
+export interface Bit {
+  postId: string,
+  dateCreated: number,
+  author: {
+    name: string,
+    personId: string,
+  },
+  upvotes: number,
+  downvotes: number,
+  title: string,
+  content: string,
+  mainChars: string[],
+  vsChars: string[],
+  stages: string[],
+  tags: string[],
+}
+
+const BitsSchema = new Schema<typeof methods>({
   postId: String,
   dateCreated: Number,
   author: {
@@ -17,5 +35,13 @@ export const BitsSchema = new Schema({
   stages: [String],
   tags: [String],
 });
+BitsSchema.statics = statics;
+BitsSchema.methods = methods;
 
-BitsSchema.statics.queryBits = queryBits;
+/** Type for a bit document. */
+export type BitDocument = Bit & Document & typeof methods;
+
+/** Type for the bit model (table, as opposed to document). */
+type IBitsModel = Model<BitDocument> & typeof statics;
+
+export const BitsModel = model<BitDocument>('Post', BitsSchema) as IBitsModel;
